feat(project): added confirmation dialog in monitor , token and project

+++ web/src/app/monitors/monitors-list/monitors-list.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
+import { MatDialog, MatDialogConfig, MatDialogRef, MatSnackBar } from '@angular/material';
 
 // Rxjs operators
 import { ActivatedRoute, Router } from '@angular/router';
@@ -6,8 +7,8 @@ import { Subscription } from 'rxjs';
 import { map } from 'rxjs/operators';
 
 // Dashboard hub application model and services
-import { MatSnackBar } from '@angular/material';
 import { MonitorService, ProjectService } from '../../core/services/index.service';
+import { DialogConfirmationComponent } from '../../shared/dialog/confirmation/dialog-confirmation.component';
 import { MonitorModel, ProjectModel } from '../../shared/models/index.model';
 
 @Component({
@@ -17,6 +18,7 @@ import { MonitorModel, ProjectModel } from '../../shared/models/index.model';
 })
 export class MonitorsListComponent implements OnInit, OnDestroy {
 
+  private dialogRef: MatDialogRef<DialogConfirmationComponent>;
   private monitorSubscription: Subscription;
   private saveMonitorSubscription: Subscription;
   public monitors: MonitorModel[] = [];
@@ -25,6 +27,7 @@ export class MonitorsListComponent implements OnInit, OnDestroy {
   public manualPing: boolean = false;
 
   constructor(
+    private dialog: MatDialog,
     private monitorService: MonitorService,
     private projectService: ProjectService,
     private route: ActivatedRoute,
@@ -57,16 +60,29 @@ export class MonitorsListComponent implements OnInit, OnDestroy {
    * @param uid the uid of monitor which needs to be deleted
    */
   deleteMonitor(monitorUid: string): void {
-    this.monitors = this.monitors.filter((monitor: MonitorModel) => monitor.uid !== monitorUid);
-    this.saveMonitorSubscription = this.monitorService.saveMonitors(this.projectUid, this.monitors)
-      .subscribe(
-        () => this.router.navigate([`/projects/${this.projectUid}/monitors`]),
-        (error: any): any => this.snackBar.open(error.message, undefined, { duration: 5000 })
-      );
+    let dialogConfig: MatDialogConfig = new MatDialogConfig();
+    dialogConfig = {
+      width: '500px',
+      data: {
+        title: 'Delete Monitor',
+        content: 'Are you sure you want to delete?',
+      },
+    };
+    this.dialogRef = this.dialog.open(DialogConfirmationComponent, dialogConfig);
+    this.dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result === true) {
+        this.monitors = this.monitors.filter((monitor: MonitorModel) => monitor.uid !== monitorUid);
+        this.saveMonitorSubscription = this.monitorService.saveMonitors(this.projectUid, this.monitors)
+          .subscribe(
+            () => this.router.navigate([`/projects/${this.projectUid}/monitors`]),
+            (error: any): any => this.snackBar.open(error.message, undefined, { duration: 5000 })
+          );
 
-    // When delete a monitor , deleting all its pings
-    this.monitorService
-      .deletePingsByMonitor(this.projectUid, monitorUid);
+        // When delete a monitor , deleting all its pings
+        this.monitorService
+          .deletePingsByMonitor(this.projectUid, monitorUid);
+      }
+    });
   }
 
   // This function will ping the monitor
